planner test skeleton
